refactor(rm): remove `delete_interface` rpc
