+++ apps/astarte_realm_management_api/lib/astarte_realm_management_api/interfaces/queries.ex
@@ -26,6 +26,7 @@ defmodule Astarte.RealmManagement.API.Interfaces.Queries do
   alias Astarte.Core.CQLUtils
   alias Astarte.DataAccess.CSystem
   alias Astarte.RealmManagement.API.CreateDatastreamIndividualMultiInterface
+  alias Astarte.Core.Device
   alias Astarte.Core.InterfaceDescriptor
   alias Astarte.DataAccess.KvStore
   alias Astarte.Core.Interface, as: InterfaceDocument
@@ -36,9 +37,12 @@ defmodule Astarte.RealmManagement.API.Interfaces.Queries do
   alias Astarte.Core.Mapping.Retention
   alias Astarte.Core.Mapping.ValueType
   alias Astarte.DataAccess.Consistency
+  alias Astarte.DataAccess.KvStore
   alias Astarte.DataAccess.Realms.Endpoint
+  alias Astarte.DataAccess.Realms.IndividualProperty
   alias Astarte.DataAccess.Realms.Interface
   alias Astarte.DataAccess.Realms.Realm
+  alias Astarte.DataAccess.Realms.SimpleTrigger
   alias Astarte.DataAccess.Repo
 
   require Logger
@@ -530,5 +534,256 @@ defmodule Astarte.RealmManagement.API.Interfaces.Queries do
         {:ok, interface_document}
       end
     end
+  end
+
+  def delete_interface(realm_name, interface_name, interface_major_version) do
+    _ =
+      Logger.info("Delete interface.",
+        interface: interface_name,
+        interface_major: interface_major_version,
+        tag: "db_delete_interface"
+      )
+
+    keyspace = Realm.keyspace_name(realm_name)
+
+    interface_id = CQLUtils.interface_id(interface_name, interface_major_version)
+
+    endpoint_query =
+      from Endpoint,
+        prefix: ^keyspace,
+        where: [interface_id: ^interface_id]
+
+    interface_query =
+      from Interface,
+        prefix: ^keyspace,
+        where: [name: ^interface_name, major_version: ^interface_major_version]
+
+    queries = [
+      Repo.to_sql(:delete_all, endpoint_query),
+      Repo.to_sql(:delete_all, interface_query)
+    ]
+
+    consistency = Consistency.domain_model(:write)
+
+    Exandra.execute_batch(
+      Repo,
+      %Exandra.Batch{
+        queries: queries
+      },
+      consistency: consistency
+    )
+  end
+
+  def delete_devices_with_data_on_interface(realm_name, interface_name) do
+    keyspace = Realm.keyspace_name(realm_name)
+    group_name = "devices-with-data-on-interface-#{interface_name}-v0"
+
+    query = from KvStore, where: [group: ^group_name]
+
+    consistency = Consistency.domain_model(:write)
+
+    _ = Repo.delete_all(query, prefix: keyspace, consistency: consistency)
+
+    :ok
+  end
+
+  def delete_interface_storage(
+        realm_name,
+        %InterfaceDescriptor{
+          storage_type: :one_object_datastream_dbtable,
+          storage: table_name
+        } = _interface_descriptor
+      ) do
+    keyspace = Realm.keyspace_name(realm_name)
+    delete_statement = "DROP TABLE IF EXISTS #{keyspace}.#{table_name}"
+    consistency = Consistency.domain_model(:write)
+
+    _ = Repo.query!(delete_statement, [], consistency: consistency)
+    _ = Logger.info("Deleted #{table_name} table.", tag: "db_delete_interface_table")
+    :ok
+  end
+
+  def delete_interface_storage(realm_name, %InterfaceDescriptor{} = interface_descriptor) do
+    with {:ok, result} <- devices_with_data_on_interface(realm_name, interface_descriptor.name) do
+      Enum.reduce_while(result, :ok, fn encoded_device_id, _acc ->
+        with {:ok, device_id} <- Device.decode_device_id(encoded_device_id),
+             :ok <- delete_values(realm_name, device_id, interface_descriptor) do
+          {:cont, :ok}
+        else
+          {:error, reason} ->
+            {:halt, {:error, reason}}
+        end
+      end)
+    end
+  end
+
+  defp delete_values(
+         realm_name,
+         device_id,
+         %InterfaceDescriptor{
+           interface_id: interface_id,
+           storage_type: :multi_interface_individual_properties_dbtable,
+           storage: table_name
+         } = _interface_descriptor
+       ) do
+    keyspace = Realm.keyspace_name(realm_name)
+
+    query =
+      from table_name,
+        where: [device_id: ^device_id, interface_id: ^interface_id]
+
+    consistency = Consistency.device_info(:write)
+
+    _ = Repo.delete_all(query, prefix: keyspace, consistency: consistency)
+
+    :ok
+  end
+
+  defp delete_values(
+         realm_name,
+         device_id,
+         %InterfaceDescriptor{
+           storage_type: :multi_interface_individual_datastream_dbtable
+         } = interface_descriptor
+       ) do
+    with {:ok, result} <-
+           fetch_all_paths_and_endpoint_ids(realm_name, device_id, interface_descriptor),
+         :ok <- delete_all_paths_values(realm_name, device_id, interface_descriptor, result) do
+      delete_all_paths(realm_name, device_id, interface_descriptor)
+    end
+  end
+
+  defp fetch_all_paths_and_endpoint_ids(
+         realm_name,
+         device_id,
+         %InterfaceDescriptor{
+           interface_id: interface_id,
+           storage_type: :multi_interface_individual_datastream_dbtable
+         } = _interface_descriptor
+       ) do
+    keyspace = Realm.keyspace_name(realm_name)
+
+    query =
+      from IndividualProperty,
+        select: [:endpoint_id, :path],
+        where: [device_id: ^device_id, interface_id: ^interface_id]
+
+    consistency = Consistency.device_info(:read)
+
+    with {:ok, properties} <- Repo.fetch_all(query, prefix: keyspace, consistency: consistency) do
+      properties =
+        Enum.map(properties, fn property ->
+          [endpoint_id: property.endpoint_id, path: property.path]
+        end)
+
+      {:ok, properties}
+    end
+  end
+
+  defp delete_all_paths(
+         realm_name,
+         device_id,
+         %InterfaceDescriptor{
+           interface_id: interface_id,
+           storage_type: :multi_interface_individual_datastream_dbtable
+         } = _interface_descriptor
+       ) do
+    keyspace = Realm.keyspace_name(realm_name)
+
+    query =
+      from IndividualProperty,
+        where: [device_id: ^device_id, interface_id: ^interface_id]
+
+    consistency = Consistency.device_info(:write)
+
+    _ = Repo.delete_all(query, consistency: consistency, prefix: keyspace)
+
+    :ok
+  end
+
+  defp delete_all_paths_values(realm_name, device_id, interface_descriptor, all_paths) do
+    Enum.reduce_while(all_paths, :ok, fn [endpoint_id: endpoint_id, path: path], _acc ->
+      with :ok <-
+             delete_path_values(realm_name, device_id, interface_descriptor, endpoint_id, path) do
+        {:cont, :ok}
+      else
+        {:error, reason} ->
+          {:halt, {:error, reason}}
+      end
+    end)
+  end
+
+  def delete_path_values(
+        realm_name,
+        device_id,
+        %InterfaceDescriptor{
+          interface_id: interface_id,
+          storage_type: :multi_interface_individual_datastream_dbtable,
+          storage: table_name
+        } = _interface_descriptor,
+        endpoint_id,
+        path
+      ) do
+    keyspace = Realm.keyspace_name(realm_name)
+
+    query =
+      from table_name,
+        where: [
+          device_id: ^device_id,
+          interface_id: ^interface_id,
+          endpoint_id: ^endpoint_id,
+          path: ^path
+        ]
+
+    consistency = Consistency.device_info(:write)
+
+    _ = Repo.delete_all(query, prefix: keyspace, consistency: consistency)
+
+    :ok
+  end
+
+  defp devices_with_data_on_interface(realm_name, interface_name) do
+    group_name = "devices-with-data-on-interface-#{interface_name}-v0"
+    keyspace = Realm.keyspace_name(realm_name)
+
+    query =
+      from map in KvStore,
+        select: map.key,
+        where: [group: ^group_name]
+
+    consistency = Consistency.domain_model(:read)
+
+    Repo.fetch_all(query, prefix: keyspace, consistency: consistency)
+  end
+
+  def has_interface_simple_triggers?(realm_name, object_id) do
+    keyspace = Realm.keyspace_name(realm_name)
+
+    simple_triggers_query =
+      from SimpleTrigger,
+        where: [object_id: ^object_id, object_type: 2]
+
+    consistency = Consistency.domain_model(:read)
+
+    # Repo.some?/2 only returns `{:ok, boolean()}`
+    {:ok, result} = Repo.some?(simple_triggers_query, prefix: keyspace, consistency: consistency)
+    result
+  end
+
+  def is_any_device_using_interface?(realm_name, interface_name) do
+    group_name = "devices-by-interface-#{interface_name}-v0"
+    keyspace = Realm.keyspace_name(realm_name)
+
+    devices_query =
+      from map in KvStore,
+        select: map.key,
+        where: [group: ^group_name],
+        limit: 1
+
+    consistency = Consistency.domain_model(:read)
+
+    # Repo.some?/2 only returns `{:ok, boolean()}`
+    {:ok, result} = Repo.some?(devices_query, prefix: keyspace, consistency: consistency)
+    result
   end
 end
