+++ apps/astarte_realm_management_api/test/support/helpers/astarte_realm_management_mock.ex
@@ -19,7 +19,6 @@
 defmodule Astarte.RealmManagement.API.Helpers.RPCMock do
   alias Astarte.RPC.Protocol.RealmManagement.{
     Call,
-    DeleteInterface,
     GetDatastreamMaximumStorageRetention,
     GetDatastreamMaximumStorageRetentionReply,
     GenericErrorReply,
@@ -55,26 +54,6 @@ defmodule Astarte.RealmManagement.API.Helpers.RPCMock do
 
   defp extract_call_tuple(%Call{call: call_tuple}) do
     call_tuple
-  end
-
-  defp execute_rpc(
-         {:delete_interface,
-          %DeleteInterface{
-            realm_name: realm_name,
-            interface_name: name,
-            interface_major_version: major,
-            async_operation: async
-          }}
-       ) do
-    case DB.delete_interface(realm_name, name, major) do
-      :ok ->
-        generic_ok(async)
-        |> ok_wrap
-
-      {:error, reason} ->
-        generic_error(reason)
-        |> ok_wrap
-    end
   end
 
   defp execute_rpc({:get_interfaces_list, %GetInterfacesList{realm_name: realm_name}}) do
